fix key access for GraphHopperWeb and POST /route
