+++ client-hc/src/test/java/com/graphhopper/api/GHMatrixSyncTest.java
@@ -22,11 +22,6 @@ public class GHMatrixSyncTest extends AbstractGHMatrixWebTester {
         return new GraphHopperMatrixWeb(new GHMatrixSyncRequester("") {
 
             @Override
-            protected String postJson(String url, JsonNode data) throws IOException {
-                return "{\"job_id\": \"1\"}";
-            }
-
-            @Override
             protected String getJson(String url) throws IOException {
                 return finalJsonStr;
             }
